rename demo-credentials.module.js back to demo-credentials.js
