+++ .gitignore
@@ -11,6 +11,4 @@ npm-debug.log
 build
 build/
 docs/
-demo-credentials.js
 **/dist/
-.hot-reload/
