+++ packages/amazon-sumerian-hosts-core/test/integration_test/core/core.lex.html
@@ -25,7 +25,7 @@
     <p>Lex Response:</p>
     <script src="https://sdk.amazonaws.com/js/aws-sdk-2.645.0.min.js"></script>
     <script type="module">
-      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.module.js';
+      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.js';
       import * as HOST from '@amazon-sumerian-hosts/core';
 
       async function main() {
