+++ packages/amazon-sumerian-hosts-core/test/integration_test/core/core.texttospeech.html
@@ -20,7 +20,7 @@
   <body>
     <script src="https://sdk.amazonaws.com/js/aws-sdk-2.645.0.min.js"></script>
     <script type="module">
-      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.module.js';
+      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.js';
       import * as HOST from '@amazon-sumerian-hosts/core';
 
       async function main() {
