+++ packages/amazon-sumerian-hosts-three/examples/three.html
@@ -186,7 +186,7 @@
       import {OrbitControls} from 'three/examples/jsm/controls/OrbitControls.js';
       import {GLTFLoader} from 'three/examples/jsm/loaders/GLTFLoader.js';
       import * as HOST from '@amazon-sumerian-hosts/three';
-      import {cognitoIdentityPoolId} from '../../../demo-credentials.module.js';
+      import {cognitoIdentityPoolId} from '../../../demo-credentials.js';
       import AWS from 'aws-sdk';
 
       const renderFn = [];
