+++ packages/amazon-sumerian-hosts-three/test/integration_test/three.js/three.texttospeech.html
@@ -34,7 +34,7 @@
       import * as THREE from 'three';
       import {OrbitControls} from 'three/examples/jsm/controls/OrbitControls.js';
       import * as HOST from '@amazon-sumerian-hosts/three';
-      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.module.js';
+      import {cognitoIdentityPoolId} from '../../../../../demo-credentials.js';
       import AWS from 'aws-sdk';
 
       async function main() {
