+++ packages/demos-babylon/src/chatbotDemo.js
@@ -2,7 +2,7 @@ import {HostObject, aws as AwsFeatures} from '@amazon-sumerian-hosts/babylon';
 import {Scene} from '@babylonjs/core/scene.js';
 import DemoUtils from './common/demo-utils.js';
 // eslint-disable-next-line
-import {cognitoIdentityPoolId} from '../../../demo-credentials.module.js';
+import {cognitoIdentityPoolId} from '../../../demo-credentials.js';
 
 let host;
 let scene;
