+++ packages/demos-babylon/src/gesturesDemo.js
@@ -2,7 +2,7 @@ import {HostObject} from '@amazon-sumerian-hosts/babylon';
 import {Scene} from '@babylonjs/core/scene.js';
 import DemoUtils from './common/demo-utils.js';
 // eslint-disable-next-line
-import {cognitoIdentityPoolId} from '../../../demo-credentials.module.js';
+import {cognitoIdentityPoolId} from '../../../demo-credentials.js';
 
 let host;
 let scene;
